new changes and improvements in implementation of repositories (not yet finished)

+++ app/src/main/java/com/example/expensetracker/domain/repository/CategoriesListRepository.kt
@@ -7,18 +7,13 @@ import kotlinx.coroutines.Dispatchers
 import kotlin.coroutines.CoroutineContext
 
 interface CategoriesListRepository {
-    suspend fun setCategoriesList(categoryDao: ExpenseCategoryDao, context: CoroutineContext = Dispatchers.IO)
+    suspend fun setCategoriesList(categoryDao: ExpenseCategoryDao, context: CoroutineContext= Dispatchers.IO)
     suspend fun addCategory(category: ExpenseCategory, context: CoroutineContext = Dispatchers.IO)
     suspend fun editCategory(category: ExpenseCategory, context: CoroutineContext = Dispatchers.IO)
     suspend fun deleteCategory(category: ExpenseCategory, context: CoroutineContext = Dispatchers.IO)
     suspend fun addDefaultCategories(context: Context, coroutineContext: CoroutineContext = Dispatchers.IO)
 
     fun getCategoriesList(): List<ExpenseCategory>
-
-    fun getCategoryItem(categoryItemId: Long): ExpenseCategory?
-
-
+    fun getCategoryItem(categoryItemId: Int): ExpenseCategory?
     fun checkDefaultCategories(context: Context): Boolean
-
-
 }
